Create lock
Even though it is so secure, if anyone enters the right number combination in the number lock, it will open up. So for safety purpose, he changes the number lock everyday. Everyday, he carries a different amount in the briefcase, and he sets the lock with a minimum possible number whose sum of digits will be equal to the money present in the briefcase.